Criação da nova fila para processar ue para remover atribuições
