+++ src/SME.SGP.Aplicacao/Interfaces/CasosDeUso/Pendencias/IRemoverAtribuicaoPendenciasUsuariosUeUseCase.cs
@@ -1,0 +1,6 @@
+﻿namespace SME.SGP.Aplicacao.Interfaces
+{
+    public interface IRemoverAtribuicaoPendenciasUsuariosUeUseCase : IRabbitUseCase
+    {
+    }
+}
